refactor: FinalFrame nextFrame 가지지 않도록 리팩토링
