+++ src/main/java/bowling/domain/Frame.java
@@ -1,13 +1,10 @@
 package bowling.domain;
 
-import java.util.Objects;
-
 public abstract class Frame {
     public static final int WAITING_CALCULATION = -1;
 
     protected Pin pin;
     protected States states;
-    protected Frame nextFrame;
 
     abstract void bowl(Pin pin);
     abstract boolean isEndFrame();
@@ -15,23 +12,7 @@ public abstract class Frame {
     abstract Frame getNextFrame(int frameNumber);
     abstract int getScore();
     public abstract States getStates();
-
-    int calculateAdditionalScore(Score score) {
-        for (Pin pin : states.getPins()) {
-            score = score.bowl(pin.getFallenPin());
-
-            if (score.canCalculateScore()) {
-                return score.getScore();
-            }
-        }
-
-        if (Objects.isNull(nextFrame)) {
-            return WAITING_CALCULATION;
-        }
-
-        return nextFrame.calculateAdditionalScore(score);
-    }
-
+    abstract int calculateAdditionalScore(Score score);
     boolean isCalculateScore() {
         return getScore() != WAITING_CALCULATION;
     }
